requirements.txt
pytorch-lightning==1.9.0
segmentation-models-pytorch==0.5.0
transformers==4.46.3
timm>=1.0.22
einops==0.8.1
opencv-python==4.11.0.86
pandas==2.2.3
albumentations==2.0.6
matplotlib>=3.9.4
tensorboard==2.19.0
pyyaml==6.0.2
ninja>=1.13.0
fvcore==0.1.5.post20221221
numpy==2.0.2
